Friends Page added and design updated
