+++ app/src/main/AndroidManifest.xml
@@ -11,8 +11,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".AddFriendActivity"></activity>
-        <activity android:name=".FriendsActivity" />
+        <activity android:name=".FriendsActivity"/>
+        <activity android:name=".AddFriendActivity"/>
+        <activity android:name=".MyFriendsActivity" />
         <activity android:name=".ProfileActivity" />
         <activity android:name=".HomePageActivity" />
         <activity android:name=".SignUpActivity" />
