+++ app/src/main/res/values/strings.xml
@@ -17,4 +17,6 @@
     <string name="search">Search</string>
     <string name="searchfriend">Search by Email</string>
     <string name="sendrequest">Send Request</string>
+    <string name="myfriends">My Friends</string>
+    <string name="friendrequests">Friend Requests</string>
 </resources>
